Transfer functions that are specific to the particle component maps
