+++ Source/Particles/WarpXParticleContainer.H
@@ -15,7 +15,6 @@
 #include "Evolve/WarpXDtType.H"
 #include "Particles/ParticleBoundaries.H"
 #include "SpeciesPhysicalProperties.H"
-#include "Utils/TextMsg.H"
 
 #ifdef WARPX_QED
 #    include "ElementaryProcess/QEDInternals/BreitWheelerEngineWrapper_fwd.H"
@@ -290,35 +289,6 @@ public:
     // split along diagonals (0) or axes (1)
     int split_type = 0;
 
-    using amrex::ParticleContainer<0, 0, PIdx::nattribs>::AddRealComp;
-    using amrex::ParticleContainer<0, 0, PIdx::nattribs>::AddIntComp;
-
-    void AddRealComp (const std::string& name, bool comm=true)
-    {
-        auto search = particle_comps.find(name);
-        if (search == particle_comps.end()) {
-            particle_comps[name] = NumRealComps();
-            particle_runtime_comps[name] = NumRealComps() - PIdx::nattribs;
-            AddRealComp(comm);
-        } else {
-            amrex::Print() << Utils::TextMsg::Info(
-                name + " already exists in particle_comps, not adding.");
-        }
-    }
-
-    void AddIntComp (const std::string& name, bool comm=true)
-    {
-        auto search = particle_icomps.find(name);
-        if (search == particle_icomps.end()) {
-            particle_icomps[name] = NumIntComps();
-            particle_runtime_icomps[name] = NumIntComps() - 0;
-            AddIntComp(comm);
-        } else {
-            amrex::Print() << Utils::TextMsg::Info(
-                name + " already exists in particle_icomps, not adding.");
-        }
-    }
-
     int doBackTransformedDiagnostics () const { return do_back_transformed_diagnostics; }
     /** Whether back-transformed diagnostics need to be performed for a particular species.
      *
@@ -327,11 +297,6 @@ public:
     void SetDoBackTransformedParticles(const bool do_back_transformed_particles) {
         m_do_back_transformed_particles = do_back_transformed_particles;
     }
-
-    std::map<std::string, int> getParticleComps () const noexcept { return particle_comps;}
-    std::map<std::string, int> getParticleiComps () const noexcept { return particle_icomps;}
-    std::map<std::string, int> getParticleRuntimeComps () const noexcept { return particle_runtime_comps;}
-    std::map<std::string, int> getParticleRuntimeiComps () const noexcept { return particle_runtime_icomps;}
 
     //amrex::Real getCharge () {return charge;}
     amrex::ParticleReal getCharge () const {return charge;}
@@ -380,11 +345,6 @@ public:
      void defineAllParticleTiles () noexcept;
 
 protected:
-    std::map<std::string, int> particle_comps;
-    std::map<std::string, int> particle_icomps;
-    std::map<std::string, int> particle_runtime_comps;
-    std::map<std::string, int> particle_runtime_icomps;
-
     int species_id;
 
     amrex::Real charge;
